feed forward implementation

+++ src/data.py
@@ -31,7 +31,7 @@ class ModelData:
         labels: A 1D numpy array containing the training labels.
     """
 
-    images: np.ndarray[Tuple[int, int, int], np.dtype[np.uint8]]
+    images: np.ndarray[Tuple[int, int, int], np.dtype[np.float64]]
     labels: np.ndarray[Tuple[int], np.dtype[np.uint8]]
 
 
@@ -77,7 +77,7 @@ def read_labels(file_name: str) -> np.ndarray[Tuple[int], np.dtype[np.uint8]]:
     return labels
 
 
-def read_images(file_name: str) -> np.ndarray[Tuple[int, int, int], np.dtype[np.uint8]]:
+def read_images(file_name: str) -> np.ndarray[Tuple[int, int, int], np.dtype[np.float64]]:
     """
     Used to read the images from the dataset file
 
@@ -92,7 +92,7 @@ def read_images(file_name: str) -> np.ndarray[Tuple[int, int, int], np.dtype[np.
         num_images = int.from_bytes(f.read(4), "big")  # "big" for big-endian
         num_rows = int.from_bytes(f.read(4), "big")  # Dimension of the image
         num_cols = int.from_bytes(f.read(4), "big")  # Dimension of the image
-        images = np.frombuffer(f.read(), dtype=np.uint8).reshape(
+        images = np.frombuffer(f.read(), dtype=np.float64).reshape(
             num_images, num_rows, num_cols
         )  # 3D array for all the images with dimensions 28x28
     return images
